Go_Marble/requirements.txt
# FastAPI framework for building APIs
fastapi==0.95.0

# ASGI server for running FastAPI applications
uvicorn==0.22.0

# Playwright for browser automation
playwright==1.30.0

# OpenAI API client for interacting with OpenAI's models
openai==0.27.0

# Python Dotenv for loading environment variables from a .env file
python-dotenv==1.0.0

# Pydantic for data validation and settings management
pydantic==1.10.2

